Update README.md
Updated the serial number submission script. Read below to learn more.

It has to be noted that the output may include illegitimate ones, usually a 8-byte substring of a 16-byte hex. This applies to apps installed from the App Store mainly, and some Developer ID apps. ASN1_PREFIX of script.sh should be amended to accommodate a 16-byte serial number. Furthermore, there's a good chance that issuerNameHash and issuerKeyHash are also different. More research is required.